Extract brush source lifecycle helpers
