+++ src/game_data_brush_source_model_internal.h
@@ -8,5 +8,7 @@ extern const char *const editor_brush_source_box_face_keys[6];
 int editor_brush_source_units_from_meters(const brush_world_runtime *world_runtime, float value);
 float editor_brush_source_meters_from_units(const brush_world_runtime *world_runtime, int value);
 bool editor_brush_source_contents_are_structural(unsigned int contents);
+void free_editor_brush_source_box(editor_brush_source_box_runtime *box);
+bool copy_source_string(char **field, const char *value);
 
 #endif
